inventory-integrations/shared_resources/app/requirements.txt
aws-lambda-powertools==2.26.0
psycopg2-binary==2.9.9
pandas==2.0.3